requirements.txt
python-telegram-bot
google-generativeai
python-dotenv
requests
beautifulsoup4
PyMuPDF